added some if statements to make sure that an enemy's position doesn't pass a waypoint by accident

+++ core/src/com/tower/defense/enemy/Enemy.java
@@ -35,14 +35,14 @@ public abstract class Enemy {
     /**
      * @return Y position
      */
-    public int getY() {
+    public float getY() {
         return posY;
     }
 
     /**
      * @return X position
      */
-    public int getX() {
+    public float getX() {
         return posX;
     }
